+++ src/components/index.ts
@@ -18,3 +18,4 @@ export * from './multi-select';
 export * from './command';
 export * from './command-cmdk';
 export * from './popover';
+export * from './range-picker';